added some javadoc for class
