+++ app/src/main/java/com/courseproject/tindar/controllers/login/LoginController.java
@@ -2,8 +2,15 @@ package com.courseproject.tindar.controllers.login;
 
 import com.courseproject.tindar.usecases.login.LoginInputBoundary;
 
+/**
+ * This class is responsible for handling user's input for authenticating and accessing login of
+ * their account
+ */
 public class LoginController {
 
+    /**
+     * Input boundary interface used for handling user's input and processing user's info
+     */
     final LoginInputBoundary userInput;
 
     public LoginController(LoginInputBoundary loginUserInput){
