+++ app/src/main/java/com/courseproject/tindar/ui/conversationlist/ConversationFragment.java
@@ -1,18 +1,17 @@
 package com.courseproject.tindar.ui.conversationlist;
 
 import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.courseproject.tindar.R;
-import com.courseproject.tindar.ui.chat.ChatActivity;
 import com.courseproject.tindar.usecases.conversationlist.ConversationResponseModel;
 
 import java.util.ArrayList;
@@ -20,7 +19,7 @@ import java.util.ArrayList;
 /**
  * A fragment representing a list of Items.
  */
-public class ConversationListFragment extends Fragment {
+public class ConversationFragment extends Fragment {
 
     // TODO: Customize parameter argument names
     private static final String ARG_COLUMN_COUNT = "column-count";
@@ -31,13 +30,13 @@ public class ConversationListFragment extends Fragment {
      * Mandatory empty constructor for the fragment manager to instantiate the
      * fragment (e.g. upon screen orientation changes).
      */
-    public ConversationListFragment() {
+    public ConversationFragment() {
     }
 
     // TODO: Customize parameter initialization
     @SuppressWarnings("unused")
-    public static ConversationListFragment newInstance(int columnCount) {
-        ConversationListFragment fragment = new ConversationListFragment();
+    public static ConversationFragment newInstance(int columnCount) {
+        ConversationFragment fragment = new ConversationFragment();
         Bundle args = new Bundle();
         args.putInt(ARG_COLUMN_COUNT, columnCount);
         fragment.setArguments(args);
@@ -61,39 +60,16 @@ public class ConversationListFragment extends Fragment {
         convos.add(new ConversationResponseModel("lisbeth", "how do you do", "20:20"));
         convos.add(new ConversationResponseModel("mikael", "silence", "20:50"));
         // Set the adapter
-
-        RecyclerView recyclerView = view.findViewById(R.id.list);
-        if (recyclerView != null) {
+        if (view instanceof RecyclerView) {
             Context context = view.getContext();
-            recyclerView.setLayoutManager(new LinearLayoutManager(context));
-
-            MyConversationRecyclerViewAdapter adapter = new MyConversationRecyclerViewAdapter(convos);
-            recyclerView.setAdapter(adapter);
-
-            // Set click listener for conversation items
-            adapter.setOnItemClickListener(conversation -> {
-                // Handle the click here, navigate to ChatActivity
-                Intent intent = new Intent(requireActivity(), ChatActivity.class);
-                intent.putExtra("conversation_partner_name", conversation.getUserName());
-                // Pass any other necessary data
-                startActivity(intent);
-            });
+            RecyclerView recyclerView = (RecyclerView) view;
+            if (mColumnCount <= 1) {
+                recyclerView.setLayoutManager(new LinearLayoutManager(context));
+            } else {
+                recyclerView.setLayoutManager(new GridLayoutManager(context, mColumnCount));
+            }
+            recyclerView.setAdapter(new MyConversationRecyclerViewAdapter(convos));
         }
-
         return view;
-    }   }
-//        RecyclerView recyclerView = null;
-//        if (view instanceof RecyclerView) {
-//            Context context = view.getContext();
-//            recyclerView = (RecyclerView) view;
-//            if (mColumnCount <= 1) {
-//                recyclerView.setLayoutManager(new LinearLayoutManager(context));
-//            } else {
-//                recyclerView.setLayoutManager(new GridLayoutManager(context, mColumnCount));
-//            }
-//            recyclerView.setAdapter(new MyConversationRecyclerViewAdapter(convos));
-//        }
-//
-//
-//        return view;
-//    }
+    }
+}
