+++ app/src/main/java/com/courseproject/tindar/ui/conversationlist/MyConversationRecyclerViewAdapter.java
@@ -9,6 +9,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.courseproject.tindar.databinding.FragmentConversationBinding;
 import com.courseproject.tindar.usecases.conversationlist.ConversationResponseModel;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -19,17 +20,13 @@ public class MyConversationRecyclerViewAdapter extends RecyclerView.Adapter<MyCo
 
     private final List<ConversationResponseModel> mValues;
 
-    public interface OnItemClickListener {
-        void onItemClick(ConversationResponseModel conversation);
-    }
-
-    private OnItemClickListener mListener;
-
-    public void setOnItemClickListener(OnItemClickListener listener) {
-        mListener = listener;
-    }
     public MyConversationRecyclerViewAdapter(List<ConversationResponseModel> items) {
         mValues = items;
+    }
+
+    public MyConversationRecyclerViewAdapter(ArrayList<ConversationResponseModel> convos) {
+
+        mValues = null;
     }
 
     @Override
@@ -41,18 +38,11 @@ public class MyConversationRecyclerViewAdapter extends RecyclerView.Adapter<MyCo
 
     @Override
     public void onBindViewHolder(final ViewHolder holder, int position) {
-        final ConversationResponseModel conversation = mValues.get(position);
-
         holder.mItem = mValues.get(position);
         holder.mUserName.setText(mValues.get(position).getUserName());
         holder.mLastMessage.setText(mValues.get(position).getLastMessage());
         holder.mLastMessageTime.setText(mValues.get(position).getLastMessageTime());
-
-        holder.itemView.setOnClickListener(v -> {
-            if (mListener != null) {
-                mListener.onItemClick(conversation);
-            }
-        });
+        
     }
 
     @Override
