+++ app/src/main/java/com/courseproject/tindar/ui/home/HomeFragment.java
@@ -4,18 +4,18 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+//import android.widget.TextView;
 import android.widget.Button;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
+//import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
 import com.courseproject.tindar.R;
 import com.courseproject.tindar.databinding.FragmentHomeBinding;
-import com.courseproject.tindar.ui.editfilters.EditFiltersFragment;
+//import com.courseproject.tindar.ui.editfilters.EditFiltersFragment;
 
 public class HomeFragment extends Fragment {
     private FragmentHomeBinding binding;
