+++ app/src/main/java/com/courseproject/tindar/usecases/conversations/ConversationListInputBoundary.java
@@ -1,9 +1,11 @@
 package com.courseproject.tindar.usecases.conversations;
 
+import com.courseproject.tindar.usecases.conversationlist.ConversationResponseModel;
+
 import java.util.ArrayList;
 
 public interface ConversationListInputBoundary {
-    // have same methods signature (implemenation will be different) as cntroller
+    // have same methods signature (implementation will be different) as controller
 
-    ArrayList<ConversationResponseModel> getActiveConversations (String userId);
+    ArrayList<ConversationResponseModel> getActiveConversationIds(String userId);
 }
